+++ inst/htmlwidgets/echarts4r.js
@@ -83,9 +83,15 @@ HTMLWidgets.widget({
             Shiny.onInputChange(el.id + '_global_out' + ":echarts4rParse", e, {priority: 'event'});
           });
 
-          chart.getZr().on("click", function(e){
-            Shiny.setInputValue(el.id + '_clicked_zr' + ":echarts4rParse", e);
-          });
+          if(x.hasOwnProperty('zr')){
+            chart.getZr().on("click", function(e){
+              delete e.stop;
+              delete e.topTarget;
+              delete e.target
+              delete e.event.path;
+              Shiny.setInputValue(el.id + '_clicked_zr' + ":echarts4rParse", e);
+            });
+          }
           
           if(x.hasOwnProperty('capture')){
             chart.on(x.capture, function(e){
@@ -201,11 +207,10 @@ HTMLWidgets.widget({
       },
 
       resize: function(width, height) {
+        if(!chart)
+          return;
 
-        if(chart){
-          chart.resize({width: width, height: height});
-        }
-
+        chart.resize({width: width, height: height});
       }
 
     };